Categorie Controller Added

+++ src/main/java/com/gamesUP/gamesUP/repository/PublisherRepository.java
@@ -4,5 +4,5 @@ import com.gamesUP.gamesUP.model.Publisher;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface PublisherRepository extends JpaRepository<Publisher,Integer> {
+public interface PublisherRepository extends JpaRepository<Publisher,Long> {
 }
